3.7.2.3 заблокировать фильтр, у которого нет подходящих задач и перенес функцию в файл mock/filter.js
