+++ src/components/filter.js
@@ -9,6 +9,7 @@ const createFilterMarkup = (filter, isChecked) => {
       class="filter__input visually-hidden"
       name="filter"
       ${isChecked ? `checked` : ``}
+      ${!count ? `disabled` : ``}
     />
     <label for="filter__${name}" class="filter__label">
       ${name} <span class="filter__${name}-count">${count}</span>
